+++ src/com/prj/mapper/exam/ExamMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ExamMapper {
 
     //查询考试
-    public List<Exam> queryExam(@Param("mid") int mid);
+    public List<Exam> queryExam(@Param("mid") long mid);
 
     //添加试题信息
     public int addExam(Exam exam);
